Recompilar nivel_alto.bin (se colgaba en Windows) y arreglar el latido del puente ntfy
sim/run.sh no compilaba en macOS/clang por un narrowing conversion de double a
float en CO2Curve[3] (los tres .ino); se agregan casts explicitos, sin cambio
de comportamiento. Con eso corriendo, el arnes encontro un bug real en
puente_serial.py: el latido que debe republicar el conteo a ntfy.sh cada 2 min
aunque no cambie nunca se disparaba, porque usaba 0.0 como centinela de "nunca
publicado" asumiendo que time.monotonic() arranca muy por encima de 5 s -- en
este Mac arranca en 0.06 s. Se cambia el centinela a None.

proyecto/nivel_alto/nivel_alto.ino compilo limpio aqui (arduino-cli 1.5.1,
esp32:esp32:esp32s3:CDCOnBoot=cdc,USBMode=default); se recompila
proyecto/placa/nivel_alto.bin, que hasta ahora tenia la build colgada de
Windows sin el comando BOOTLOADER.

sim/run.sh: 0 fallos (antes: 1, el del latido).

+++ proyecto/nivel_medio/nivel_medio.ino
@@ -53,7 +53,7 @@
 const float DC_GAIN = 8.5;
 const float ZERO_POINT_VOLTAGE = 0.265;
 const float REACTION_VOLTAGE = 0.059;
-const float CO2Curve[3] = {2.602, ZERO_POINT_VOLTAGE, (REACTION_VOLTAGE / (2.602 - 3))};
+const float CO2Curve[3] = {2.602f, ZERO_POINT_VOLTAGE, (float)(REACTION_VOLTAGE / (2.602 - 3))};
 
 LiquidCrystal_I2C lcd(0x27, 20, 4);
